Add files via upload
The code that takes the RSS feed URL and the path to save the images as command-line arguments:

This code first checks if there are at least two command-line arguments (sys.argv is a list that contains the command-line arguments passed to the Python script). If there are fewer than two arguments, it prints a usage message and exits the script.

It then assigns the first command-line argument (sys.argv[1]) to rss_url and the second command-line argument (sys.argv[2]) to save_dir. This allows you to specify the RSS feed URL and the path to save the images when you run the script from the command line, like this:

example :- 
```
python download_images.py https://example.com/feed /path/to/images/directory

```